+++ plugins/dates/src/durations/index.js
@@ -73,6 +73,10 @@ const addDurations = function (Doc) {
     let m = this.match('#Value+ #Duration (and? #Value+ #Duration)?')
     // add '20mins'
     m = m.concat(this.match('(#Duration && /[0-9][a-z]+$/)'))
+
+    // not 'in 20 minutes'
+    m = m.notIf('#DateShift')
+
     if (typeof n === 'number') {
       m = m.get(n)
     }